Require time lib so util method can convert to Time objects (#38)
In the `Util.deep_parse_iso8601_values` we reference `Time.iso8601` but
to use that in Ruby you need to require the time library otherwise the
method is undefined. The implementation has a `rescue` but that was
unintended since its only meant to capture cases where a string can’t
be converted to a time object. We weren’t seeing any errors here
because it also caught the undefined error and let the time be a string
value.